+++ app/controllers/admin/services_controller.rb
@@ -1,12 +1,15 @@
 class Admin
   class ServicesController < ApplicationController
     include Taggable
+    include Searchable
 
     before_action :authenticate_admin!
     layout 'admin'
 
     def index
-      @services = Kaminari.paginate_array(policy_scope(Service)).
+      @search_term = search_params(params)[:q]
+      all_services = search(policy_scope(Service), @search_term, 2)
+      @services = Kaminari.paginate_array(all_services).
                   page(params[:page]).per(params[:per_page])
     end